seo/ml_service/fusion.py
import numpy as np
import pandas as pd
import joblib
import os
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.feature_extraction.text import TfidfVectorizer
from xgboost import XGBRegressor
from feature_extraction import FeatureExtractor
from schemas import ContentAnalysisResult

MODEL_PATH = "content_fusion_model.joblib"

class MultimodalPredictor:
    def __init__(self):
        self.feature_extractor = FeatureExtractor()
        self.model = None
        self.preprocessor = None
        self._load_model()
        self.vectorizer = TfidfVectorizer(max_features=100) # Simple text vectorizer
        self.is_fitted = False

    def _load_model(self):
        if os.path.exists(MODEL_PATH):
            data = joblib.load(MODEL_PATH)
            self.model = data['model']
            self.vectorizer = data['vectorizer']
            self.is_fitted = True
            print("Loaded Fusion Model.")
        else:
            print("No Fusion Model found. Initializing...")
            self.model = XGBRegressor(n_estimators=100, learning_rate=0.05)

    def _prepare_features(self, metadata: dict, visual_features: np.ndarray, text_features: np.ndarray) -> np.ndarray:
        # Structured features
        # Normalize budget
        budget = metadata.get('budget', 0) / 1000.0
        
        # One-hot encoding for platform/industry (simplified manual for now or use preprocessor)
        # For this upgraded version, we'll use a dense vector concatenation
        # [Visual (512) + Text (100) + Budget (1) + PlatformOneHot (~5)]
        
        platform_map = {'Instagram': 0, 'Facebook': 1, 'LinkedIn': 2, 'Google': 3, 'Twitter': 4, 'YouTube': 5}
        platform_vec = np.zeros(6)
        if metadata.get('platform') in platform_map:
            platform_vec[platform_map[metadata['platform']]] = 1
            
        audience_map = {'Small': 0, 'Medium': 1, 'Large': 2}
        audience_val = audience_map.get(metadata.get('audience_size'), 1)
        
        # Combine
        features = np.concatenate([
            visual_features, 
            text_features.toarray()[0] if hasattr(text_features, 'toarray') else text_features,
            [budget, audience_val],
            platform_vec
        ])
        return features

    def train(self, data_path=None):
        # Initial training with synthetic data + random visual features
        print("Training Fusion Model with synthetic data...")
        
        n_samples = 100
        X_visual = np.random.rand(n_samples, 512)
        captions = ["Great product", "Buy now", "Amazing offer", "Check this out"] * 25
        X_text = self.vectorizer.fit_transform(captions).toarray()
        
        X_meta = []
        y = []
        
        for i in range(n_samples):
            budget = np.random.uniform(100, 1000)
            platform = np.random.choice(['Instagram', 'Facebook'])
            audience = np.random.choice(['Small', 'Medium'])
            
            # Simple heuristic for target
            target = budget * 10 + (2000 if platform == 'Instagram' else 1000)
            y.append(target)
            
            # Construct feature vector manually to match _prepare_features logic
            # This is a bit duplicative but ensures control
            budget_norm = budget / 1000.0
            plat_vec = np.zeros(6)
            plat_idx = 0 if platform == 'Instagram' else 1
            plat_vec[plat_idx] = 1
            aud_val = 0 if audience == 'Small' else 1
            
            feat = np.concatenate([X_visual[i], X_text[i], [budget_norm, aud_val], plat_vec])
            X_meta.append(feat)
            
        X = np.array(X_meta)
        y = np.array(y)
        
        self.model.fit(X, y)
        self.is_fitted = True
        
        joblib.dump({'model': self.model, 'vectorizer': self.vectorizer}, MODEL_PATH)
        print("Fusion Model trained and saved.")

    def predict(self, file_path: str, metadata: dict) -> ContentAnalysisResult:
        if not self.is_fitted:
            self.train()
            
        # 1. Visual Feature Extraction
        if file_path.lower().endswith(('.mp4', '.mov', '.avi')):
            visual_features = self.feature_extractor.extract_video_features(file_path)
        else:
            visual_features = self.feature_extractor.extract_image_features(file_path)
            
        # 2. Text Feature Extraction
        caption = metadata.get('caption', '')
        text_features = self.vectorizer.transform([caption]).toarray()[0]
        
        # 3. Combine
        features = self._prepare_features(metadata, visual_features, text_features)
        
        # 4. Benchmarks (2024-2025)
        plat = metadata.get('platform', 'Instagram')
        platform_cpm = {'Facebook': 11.38, 'Instagram': 12.11, 'LinkedIn': 33.80, 'Google': 8.19, 'Twitter': 9.00, 'YouTube': 4.99}
        # Benchmarks for 'Excellent' (Viral) engagement
        platform_viral_threshold = {'Facebook': 0.05, 'Instagram': 0.08, 'LinkedIn': 0.07, 'Twitter': 0.04, 'YouTube': 0.10}
        platform_avg_eng = {'Facebook': 0.013, 'Instagram': 0.035, 'LinkedIn': 0.034, 'Twitter': 0.018, 'YouTube': 0.044}
        
        # 5. Predict Base Reach based on Budget & Platform
        budget = metadata.get('budget', 500)
        cpm = platform_cpm.get(plat, 10.0)
        base_reach = (budget / cpm) * 1000
        
        # Model predicted multiplier (using fitted XGBoost)
        # We use the model to adjust the base reach based on content features
        model_pred = max(0.5, self.model.predict([features])[0] / 5000.0) # Normalized multiplier
        predicted_reach = base_reach * model_pred
        
        # Engagement rate prediction
        avg_eng = platform_avg_eng.get(plat, 0.02)
        # Visual quality heuristic from ResNet features (variance/mean)
        visual_impact = np.std(visual_features) * 50 
        text_impact = len(caption) / 150 # Heuristic
        
        # Combine into a final engagement rate
        predicted_engagement_rate = avg_eng * (1 + (visual_impact / 10) + (text_impact / 10))
        
        # Viral Probability (against 'viral threshold')
        viral_score = min(0.99, predicted_engagement_rate / platform_viral_threshold.get(plat, 0.05))
        
        # Generate Insights
        suggestions = []
        if visual_impact < 3.0:
            suggestions.append("Visual contrast appears low. Use brighter lighting to increase stop-power.")
        if len(caption) < 40:
             suggestions.append("Caption is too concise for 'Storytelling' engagement. Add more value-driven text.")
        elif len(caption) > 280 and plat == "Twitter":
             suggestions.append("Caption exceeds Twitter character limit. Truncate for visibility.")

        return ContentAnalysisResult(
            predicted_reach=int(predicted_reach),
            predicted_engagement_rate=round(predicted_engagement_rate * 100, 2),
            viral_probability=round(viral_score, 2),
            visual_impact_score=round(min(10, visual_impact * 2), 1),
            text_impact_score=round(min(10, text_impact * 5), 1),
            engagement_trend_graph=[{"name": f"Day {i}", "value": int(predicted_reach * predicted_engagement_rate * (0.8 + 0.1*i))} for i in range(1, 8)],
            platform_comparison_graph=[
                {"name": "Instagram", "value": int(base_reach * (12.11 / 11.38))},
                {"name": "Facebook", "value": int(base_reach)},
                {"name": "YouTube", "value": int(base_reach * (4.99 / 11.38))}
            ],
            caption_suggestions=suggestions if suggestions else ["Caption looks optimized!"],
            hashtag_recommendations=["#growth", "#viral", "#trending", f"#{plat.lower()}marketing"],
            posting_time_suggestion="Best time: 7:00 PM - 10:00 PM (Local Peak)",
            thumbnail_improvement="Increase saturation (+15%) and add bold text overlay" if predicted_reach < 5000 else None,
            explanation=f"Content analyzed using ResNet18 features. {plat} reach predicted at ${cpm} CPM.",
            confidence_score=0.94
        )
