seo/ml_service/requirements.txt
fastapi
uvicorn
scikit-learn
xgboost
pandas
numpy
joblib
pydantic
python-multipart
torch
torchvision
opencv-python-headless
pillow
vaderSentiment
textblob
nltk
transformers
detoxify
keybert
nrclex
matplotlib
seaborn
wordcloud
tqdm
openpyxl
youtube-comment-downloader
